@@ -16,7 +16,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.0.6.RELEASE</version>
+		<version>2.6.2</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	
@@ -72,9 +72,9 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<java.version>1.8</java.version>
+		<java.version>11</java.version>
 		
-		<mixednutz-app.version>2.0.18</mixednutz-app.version> 
+		<mixednutz-app.version>2.1.0-SNAPSHOT</mixednutz-app.version> 
 		<spring-social.version>1.1.6.RELEASE</spring-social.version>
 		<epublib.version>3.1</epublib.version>
 		<version.rome>1.8.0</version.rome>
@@ -97,22 +97,26 @@
 			
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
+				<artifactId>maven-surefire-plugin</artifactId> 
 				<configuration>
-					<excludedGroups>net.mixednutz.api.core.test.IntegrationTest</excludedGroups>
-				</configuration>
-			</plugin>
-	
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-failsafe-plugin</artifactId>
-				<configuration>
-					<includes>
-						<include>**/*.java</include>
-					</includes>
-					<groups>net.mixednutz.api.core.test.IntegrationTest</groups>
-				</configuration>
-			</plugin>
+					<properties>
+	                    <excludeTags>IntegrationTest</excludeTags>
+	                </properties>
+    			</configuration>
+			 </plugin>
+			 
+			 <plugin>
+			 	<groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <configuration>
+                	<includes>
+				      <include>**/*.java</include>
+				    </includes>
+                    <properties>
+	                    <includeTags>IntegrationTest</includeTags>
+	                </properties>
+                </configuration>
+             </plugin>
 	
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
@@ -218,11 +222,6 @@
 		</dependency>
 		
 		<!-- Testing -->
-	  	<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<scope>test</scope>
-		</dependency>
 		<dependency>
 			<groupId>org.hsqldb</groupId>
 			<artifactId>hsqldb</artifactId>
