@@ -1,8 +1,8 @@
 package net.mixednutz.app.server.manager.post.series.impl;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import net.mixednutz.app.server.entity.post.series.Series;
 
